feat(norm): separate action creators and thunks
